requirements.txt
mecab-python3
discord.py
